RavenDB-20238 refactor
